# apps/frontend/src/assets/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "企业级 Go+Gin+Gorm Web 后端项目模板",
    "title": "Gin Admin Starter API",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "name": "API Support",
      "url": "http://www.swagger.io/support",
      "email": "support@swagger.io"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "description": "用户名密码登录，返回 JWT Token、登录时间和过期时间",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["认证管理"],
        "summary": "用户登录",
        "parameters": [
          {
            "description": "登录信息",
            "name": "loginReq",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "password": {
                  "type": "string"
                },
                "username": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "expire_at": {
                          "type": "string"
                        },
                        "login_time": {
                          "type": "string"
                        },
                        "token": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "登出当前用户，将 Token 加入黑名单",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["认证管理"],
        "summary": "用户登出",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
            }
          }
        }
      }
    },
...
    "/api/v1/auth/register": {
      "post": {
        "description": "注册新用户账号",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["认证管理"],
        "summary": "用户注册",
        "parameters": [
          {
            "description": "用户信息",
            "name": "user",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/go-admin-starter_internal_model.User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/go-admin-starter_internal_model.User"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
            }
          }
        }
      }
    },
...
    "/api/v1/files": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "分页获取文件列表，支持按business_id或user_id筛选",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["文件管理"],
        "summary": "获取文件列表",
        "parameters": [
          {
            "type": "integer",
            "default": 1,
            "description": "页码",
            "name": "page",
            "in": "query"
          },
          {
            "type": "integer",
            "default": 10,
            "description": "每页数量",
            "name": "pageSize",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "业务ID",
            "name": "business_id",
            "in": "query"
          },
          {
            "type": "string",
            "description": "业务类型",
            "name": "business_type",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "用户ID",
            "name": "user_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "$ref": "#/definitions/go-admin-starter_internal_model.File"
                          }
                        },
                        "page": {
                          "type": "integer"
                        },
                        "pageSize": {
                          "type": "integer"
                        },
                        "total": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
            }
          }
        }
      }
    },
...
    "/api/v1/files/upload": {
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "上传单个文件",
        "consumes": ["multipart/form-data"],
        "produces": ["application/json"],
        "tags": ["文件管理"],
        "summary": "上传文件",
        "parameters": [
          {
            "type": "file",
            "description": "文件",
            "name": "file",
            "in": "formData",
            "required": true
          },
          {
            "type": "integer",
            "description": "业务ID",
            "name": "business_id",
            "in": "query"
          },
          {
            "type": "string",
            "description": "业务类型",
            "name": "business_type",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/go-admin-starter_internal_model.File"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
            }
          }
        }
      }
    },
    "/api/v1/files/upload/multiple": {
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "上传多个文件",
        "consumes": ["multipart/form-data"],
        "produces": ["application/json"],
        "tags": ["文件管理"],
        "summary": "批量上传文件",
        "parameters": [
          {
            "type": "file",
            "description": "文件列表",
            "name": "files",
            "in": "formData",
            "required": true
          },
          {
            "type": "integer",
            "description": "业务ID",
            "name": "business_id",
            "in": "query"
          },
          {
            "type": "string",
            "description": "业务类型",
            "name": "business_type",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/definitions/go-admin-starter_internal_model.File"
                      }
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
            }
          }
        }
      }
    },
    "/api/v1/files/{id}": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "根据ID获取文件详情",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["文件管理"],
        "summary": "获取文件详情",
        "parameters": [
          {
            "type": "integer",
            "description": "文件ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/go-admin-starter_internal_model.File"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
            }
          }
        }
      },
      "delete": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "删除指定文件",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["文件管理"],
        "summary": "删除文件",
        "parameters": [
          {
            "type": "integer",
            "description": "文件ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "go-admin-starter_internal_model.File": {
      "type": "object",
      "properties": {
        "businessId": {
          "type": "integer"
        },
        "businessType": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "createdBy": {
          "type": "integer"
        },
        "deletedAt": {
          "$ref": "#/definitions/gorm.DeletedAt"
        },
        "extension": {
          "type": "string"
        },
        "fileName": {
          "type": "string"
        },
        "filePath": {
          "type": "string"
        },
        "fileSize": {
          "type": "integer"
        },
        "id": {
          "type": "integer"
        },
        "mimeType": {
          "type": "string"
        },
        "originalName": {
          "type": "string"
        },
        "status": {
          "type": "integer"
        },
        "updatedAt": {
          "type": "string"
        },
        "updatedBy": {
          "type": "integer"
        }
      }
    },
...
    "go-admin-starter_internal_model.User": {
      "type": "object",
      "required": ["username"],
      "properties": {
        "createdAt": {
          "type": "string"
        },
        "createdBy": {
          "type": "integer"
        },
        "deletedAt": {
          "$ref": "#/definitions/gorm.DeletedAt"
        },
        "email": {
          "description": "邮箱，唯一索引",
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "nickname": {
          "description": "昵称",
          "type": "string"
        },
        "phone": {
          "description": "手机号",
          "type": "string"
        },
        "status": {
          "description": "状态，1: 正常, 2: 禁用",
          "type": "integer"
        },
        "updatedAt": {
          "type": "string"
        },
        "updatedBy": {
          "type": "integer"
        },
        "username": {
          "description": "用户名，唯一索引",
          "type": "string"
        }
      }
    },
...
    "go-admin-starter_pkg_response.Response": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer"
        },
        "data": {},
        "message": {
          "type": "string"
        }
      }
    },
    "gorm.DeletedAt": {
      "type": "object",
      "properties": {
        "time": {
          "type": "string"
        },
        "valid": {
          "description": "Valid is true if Time is not NULL",
          "type": "boolean"
        }
      }
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "description": "请输入 JWT Token，格式: Bearer {token}",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}